save codel semain3

+++ codel/src/domain/PhoneNumber.java
@@ -2,7 +2,7 @@ package domain;
 
 public class PhoneNumber
 {
-	protected int id;
+	protected long id;
 	protected String phoneKind;
 	protected String phoneNumber;
 	protected Contact contact;
@@ -11,12 +11,12 @@ public class PhoneNumber
 	{
 	}
 
-	public int getId()
+	public long getId()
 	{
 		return id;
 	}
 
-	public void setId(int id)
+	public void setId(long id)
 	{
 		this.id = id;
 	}
